change : the schema

+++ migrations/sqls/20221031190652-products-table-up.sql
@@ -1,7 +1,7 @@
 CREATE TABLE Products(
     id SERIAL PRIMARY  KEY,
     name VARCHAR(200) NOT NULL,
-    price INTEGER,
+    price NUMERIC,
     category VARCHAR(100),
     description TEXT
 );
